Passive Seperation
- Moves methods to their respective passive classes.
- Renames Hydrosink,java to HydroSink.java
