+++ src/com/projectkorra/projectkorra/chiblocking/passive/Acrobatics.java
@@ -5,11 +5,16 @@ import org.bukkit.entity.Player;
 
 import com.projectkorra.projectkorra.ability.ChiAbility;
 import com.projectkorra.projectkorra.ability.PassiveAbility;
+import com.projectkorra.projectkorra.configuration.ConfigManager;
 
 public class Acrobatics extends ChiAbility implements PassiveAbility {
 
 	public Acrobatics(Player player) {
 		super(player);
+	}
+
+	public static double getFallReductionFactor() {
+		return ConfigManager.getConfig().getDouble("Abilities.Chi.Passive.Acrobatics.FallReductionFactor");
 	}
 
 	@Override
