+++ src/com/projectkorra/projectkorra/chiblocking/passive/ChiPassive.java
@@ -71,23 +71,7 @@ public class ChiPassive {
 			}
 		}.runTaskTimer(ProjectKorra.plugin, 0, 1);
 	}
-
-	public static double getExhaustionFactor() {
-		return ConfigManager.getConfig().getDouble("Abilities.Chi.Passive.ChiSaturation.ExhaustionFactor");
-	}
-
-	public static double getFallReductionFactor() {
-		return ConfigManager.getConfig().getDouble("Abilities.Chi.Passive.Acrobatics.FallReductionFactor");
-	}
-
-	public static int getJumpPower() {
-		return ConfigManager.getConfig().getInt("Abilities.Chi.Passive.ChiAgility.JumpPower");
-	}
-
-	public static int getSpeedPower() {
-		return ConfigManager.getConfig().getInt("Abilities.Chi.Passive.ChiAgility.SpeedPower");
-	}
-
+	
 	public static double getChance() {
 		return ConfigManager.getConfig().getDouble("Abilities.Chi.Passive.BlockChi.Chance");
 	}
